Fix integer overflow for timer wakeup
esp_sleep_enable_timer_wakeup takes an uint64_t which is the duration in microseconds. I was using a long to convert seconds to microseconds, which would overflow if the duration is larger than 35m47s.

+++ sleep_bins.cpp
@@ -23,7 +23,7 @@ void sleep_until_5pm() {
 
   t_now = mktime(&now);
 
-  esp_sleep_enable_timer_wakeup((t_next - t_now) * 1000000L);
+  esp_sleep_enable_timer_wakeup((uint64_t)(t_next - t_now) * 1000000ULL);
   touchSleepWakeUpEnable(T0, 40);
   esp_sleep_enable_touchpad_wakeup();
   Serial.println("Going to sleep now");
